// src/app/interfaces/usuario-interface.ts
// Generated by https://quicktype.io

export interface ResultUsuario {
    ok:   boolean;
    msg:  string;
    resp: Resp[];
}

export interface Resp {
    id:       number;
    nombre:   string;
    apellido: string;
    usuario:  string;
    password: string;
    id_rol:   number;
    estado:   number;
    Rol: Rol;
}

export interface Rol {
    id: number;
    tipo: string;
    estado: number;
}